pyproject.toml
[tool.poetry]
name = "postgres-aws-s3"
version = "0.1.0"
description = ""
authors = ["Janne Keskitalo <jke@collectivecrunch.com>"]
readme = "README.md"
packages = [{include = "postgres_aws_s3"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
testcontainers = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
